+++ src/commands/tuntap.constants.ts
@@ -1,0 +1,8 @@
+/**
+ * Tunnel available modes.
+ * @category Constants
+ */
+export enum TunTapTunnelModes {
+  Tun = 'tun',
+  Tap = 'tap',
+}